+++ src/Classes/PackageUtil.php
@@ -1,0 +1,46 @@
+<?php
+
+declare(strict_types=1);
+
+/**
+ * Contao Utilities for Contao Open Source CMS
+ * Copyright (c) 2019-2023 Web ex Machina
+ *
+ * @category ContaoBundle
+ * @package  Web-Ex-Machina/contao-utils
+ * @author   Web ex Machina <contact@webexmachina.fr>
+ * @link     https://github.com/Web-Ex-Machina/contao-utils/
+ */
+
+namespace WEM\UtilsBundle\Classes;
+
+use Contao\System;
+
+class PackageUtil
+{
+    /**
+     * Get a package's version.
+     *
+     * @param string $package The package name
+     *
+     * @return string|null The package version if found, null otherwise
+     */
+    public static function getVersion(
+        string $package
+    ): string|null {
+        $packages = json_decode(
+            file_get_contents(System::getContainer()->getParameter(
+                'kernel.project_dir'
+            ) . '/vendor/composer/installed.json')
+        );
+
+        foreach ($packages->packages as $p) {
+            $p = (array) $p;
+            if ($package === $p['name']) {
+                return $p['version'];
+            }
+        }
+
+        return null;
+    }
+}