+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 Extension "Utilities" for Contao Open Source CMS
 ========
 
+2.5.0 - 2026-08-27
+- Feat : Add new `WEM\UtilsBundle\Classes\PackageUtil` class
+- Feat : Add new `WEM\UtilsBundle\Classes\PackageUtil::getVersion()` function that return the package version installed
+
 2.4.0 - 2026-08-24
 - Feat : Add new `WEM\UtilsBundle\Classes\PdfUtil` class
 - Feat : Add new `WEM\UtilsBundle\Classes\PdfUtil::getPageAsJpeg()` function to extract a page from PDF and return it as a picture path